+++ noodles-bam/src/writer.rs
@@ -293,7 +293,8 @@ where
     writer.write_u32::<LittleEndian>(l_name)?;
     writer.write_all(name)?;
 
-    let l_ref = reference_sequence.len();
+    let l_ref = i32::try_from(usize::from(reference_sequence.len()))
+        .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_i32::<LittleEndian>(l_ref)?;
 
     Ok(())
@@ -334,7 +335,7 @@ mod tests {
 
         let reference_sequences = [("sq0".parse()?, 8)]
             .into_iter()
-            .map(|(name, len): (reference_sequence::Name, i32)| {
+            .map(|(name, len): (reference_sequence::Name, usize)| {
                 let sn = name.to_string();
                 ReferenceSequence::new(name, len).map(|rs| (sn, rs))
             })