+++ noodles-sam/src/header/builder.rs
@@ -55,7 +55,7 @@ impl Builder {
     ///
     /// let reference_sequences = [("sq0".parse()?, 13)]
     ///     .into_iter()
-    ///     .map(|(name, len): (reference_sequence::Name, i32)| {
+    ///     .map(|(name, len): (reference_sequence::Name, usize)| {
     ///         let sn = name.to_string();
     ///         ReferenceSequence::new(name, len).map(|rs| (sn, rs))
     ///     })